LPS-140152 Fixed graphql

+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/common/services/liferay/graphql/queries.js
@@ -325,7 +325,7 @@ export const getListTypeDefinitions = gql`
 `;
 
 export const getAccounts = gql`
-	query getAccounts($pageSize: Long) {
+	query getAccounts($pageSize: Int) {
 		accounts(pageSize: $pageSize) {
 			items {
 				externalReferenceCode
